Added events into database
I have added the rest of the event pages and info into the database but there are some that don't need images/sponsors so there are just image symbols in the way. I also tried to figure out how to fix the footer. In the end I couldn't move the footer into the middle of the screen.